merge: QuestList クラスのメソッド整備と quest-definition.cpp/h 分離（hengband#5431 部分相当）
変愚蛮怒 PR #5431 (Improve-QuestList) のうち、QuestList API 整備と定義分離を
bakabakaband 構造に適応してマージ:
- QuestList にアクセサメソッド (set_defeated_monster / set_max_monster /
  set_type / set_monrace_id / set_flags / is_quest_equals / is_bounty_valid /
  get_sorted_quest_ids) を整備し QuestType への直接アクセスを低減 (上流 57cd183)
- QuestType::reset() / QuestList::reset_all() を定義 (上流 8cfb4ae)
- QuestId / QuestStatusType / QuestKindType / QuestType / QuestList を
  quest.cpp/h から system/dungeon/quest-definition.cpp/h に分離 (上流 b194645)

bakabakaband 独自適応:
- QuestType の get/set 系シグネチャは baka の型 (CreatureEntity / ArtifactType /
  BIT_FLAGS / 各種エイリアス) を維持。baka 固有の get_reward_artifact() および
  MELKO クエスト ID も保持。
- quest.h は後方互換のため quest-definition.h を re-include し、既存の全
  呼出元 (28 ファイル) の include を不変に保つ。
- game-play-initializer の quest リセットは reset_all() を採用しつつ、baka で
  廃止済みの inven_cnt / equip_cnt 代入は除外。

不採用 (baka の方針/差異):
- parse_quest_info() 内部整理 / create_ifs_quest_list 分離 / quest-list.h への
  再分割 / angband_fgets→ifstream (上流 f17ad95, 5c4524e, 2c69696, 16043b7) は
  baka のクエストファイルパース処理 (バリアント別処理を含む) と乖離が大きく、
  純粋な内部整理のため不採用。

上流コミット: 57cd183, 8cfb4ae, b194645 (hengband/develop)
