+++ src/Makefile.am
@@ -1380,6 +1380,7 @@ hengband_SOURCES = \
 	system/dungeon/dungeon-definition.cpp system/dungeon/dungeon-definition.h \
 	system/dungeon/dungeon-list.cpp system/dungeon/dungeon-list.h \
 	system/dungeon/dungeon-record.cpp system/dungeon/dungeon-record.h \
+	system/dungeon/quest-definition.cpp system/dungeon/quest-definition.h \
 	\
 	system/enums/grid-flow.h \
 	system/enums/grid-count-kind.h \
