+++ src/core/game-play.cpp
@@ -75,6 +75,7 @@
 #include "system/angband-system.h"
 #include "system/angband-version.h"
 #include "system/creature-entity.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/floor/floor-info.h"
 #include "system/floor/wilderness-grid.h"
