+++ src/dungeon/quest.cpp
@@ -6,7 +6,6 @@
 #include "floor/floor-mode-changer.h"
 #include "floor/floor-object.h"
 #include "game-option/play-record-options.h"
-#include "info-reader/fixed-map-parser.h"
 #include "io/write-diary.h"
 #include "locale/english.h"
 #include "main/music-definitions-table.h"
@@ -25,6 +24,7 @@
 #include "system/artifact-type-definition.h"
 #include "system/creature-entity.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/floor/floor-info.h" // @todo 相互参照、将来的に削除する.
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
@@ -37,9 +37,6 @@
 #include <sstream>
 #include <stdexcept>
 
-std::vector<std::string> quest_text_lines; /*!< Quest text */
-QuestId leaving_quest = QuestId::NONE;
-
 /*!
  * @brief クエスト突入時のメッセージテーブル
  */
@@ -51,159 +48,6 @@ const std::vector<std::string> quest_entered_messages = {
     _("何かが階段の上に書いてある:", "Something is written on the staircase"),
     _("巻物を見つけた。メッセージが書いてある:", "You find a scroll with the following message"),
 };
-}
-
-/*!
- * @brief 該当IDが固定クエストかどうかを判定する.
- * @param quest_id クエストID
- * @return 固定クエストならばTRUEを返す
- */
-bool QuestType::is_fixed(QuestId quest_id)
-{
-    return (enum2i(quest_id) < MIN_RANDOM_QUEST) || (enum2i(quest_id) > MAX_RANDOM_QUEST);
-}
-
-bool QuestType::has_reward() const
-{
-    return this->reward_fa_id.has_value();
-}
-
-tl::optional<FixedArtifactId> QuestType::get_reward() const
-{
-    return this->reward_fa_id;
-}
-
-ArtifactType &QuestType::get_reward_artifact() const
-{
-    auto &artifacts = ArtifactList::get_instance();
-    return artifacts.get_artifact(this->reward_fa_id.value_or(FixedArtifactId::NONE));
-}
-
-short QuestType::get_reward_bi_id() const
-{
-    if (!this->has_reward()) {
-        return 0;
-    }
-
-    const auto &artifact = ArtifactList::get_instance().get_artifact(*this->reward_fa_id);
-    return BaseitemList::get_instance().lookup_baseitem_id(artifact.bi_key);
-}
-
-bool QuestType::is_reward_instant_artifact() const
-{
-    if (!this->has_reward()) {
-        return false;
-    }
-
-    return ArtifactList::get_instance().get_artifact(*this->reward_fa_id).gen_flags.has(ItemGenerationTraitType::INSTA_ART);
-}
-
-bool QuestType::is_reward_target(const BaseitemKey &key) const
-{
-    if (!this->has_reward()) {
-        return false;
-    }
-
-    return ArtifactList::get_instance().get_artifact(*this->reward_fa_id).bi_key == key;
-}
-
-void QuestType::set_reward(FixedArtifactId fa_id)
-{
-    if (fa_id == FixedArtifactId::NONE) {
-        this->reset_reward();
-        return;
-    }
-
-    // 馬鹿馬鹿固有: 上流の ArtifactRecords は未導入のため、
-    // 既存通り ArtifactType::gen_flags の QUESTITEM ビットでクエスト報酬状態を管理する。
-    auto &artifacts = ArtifactList::get_instance();
-    if (this->reward_fa_id && (*this->reward_fa_id != fa_id)) {
-        artifacts.get_artifact(*this->reward_fa_id).gen_flags.reset(ItemGenerationTraitType::QUESTITEM);
-    }
-
-    this->reward_fa_id = fa_id;
-    artifacts.get_artifact(fa_id).gen_flags.set(ItemGenerationTraitType::QUESTITEM);
-}
-
-void QuestType::reset_reward()
-{
-    if (this->reward_fa_id) {
-        ArtifactList::get_instance().get_artifact(*this->reward_fa_id).gen_flags.reset(ItemGenerationTraitType::QUESTITEM);
-        this->reward_fa_id.reset();
-    }
-}
-
-/*!
- * @brief 討伐対象モンスターを返す. いなければプレイヤー (無効値の意)
- * @return 討伐対象モンスター
- */
-MonraceDefinition &QuestType::get_bounty()
-{
-    return MonraceList::get_instance().get_monrace(this->r_idx);
-}
-
-/*!
- * @brief 討伐対象モンスターを返す. いなければプレイヤー (無効値の意)
- * @return 討伐対象モンスター
- */
-const MonraceDefinition &QuestType::get_bounty() const
-{
-    return MonraceList::get_instance().get_monrace(this->r_idx);
-}
-
-QuestList QuestList::instance{};
-
-QuestList &QuestList::get_instance()
-{
-    return instance;
-}
-
-/*!
- * @brief クエストの初期化
- * @details ソフトウェア起動時ではパース関数が動作しないので、各種初期化シーケンス時に遅延初期化する
- */
-void QuestList::initialize()
-{
-    try {
-        const auto quest_numbers = parse_quest_info(QUEST_DEFINITION_LIST);
-        QuestType quest{};
-        quest.status = QuestStatusType::UNTAKEN;
-        this->quests.emplace(QuestId::NONE, quest);
-        for (const auto q : quest_numbers) {
-            this->quests.emplace(q, quest);
-        }
-    } catch (const std::runtime_error &r) {
-        std::stringstream ss;
-        ss << _("ファイル読み込みエラー: ", "File loading error: ") << r.what();
-        msg_print(ss.str());
-        msg_erase();
-        quit(_("クエスト初期化エラー", "Error of quests initializing"));
-    }
-}
-
-QuestType &QuestList::get_quest(QuestId id)
-{
-    return this->quests.at(id);
-}
-
-const QuestType &QuestList::get_quest(QuestId id) const
-{
-    return this->quests.at(id);
-}
-
-std::vector<QuestId> QuestList::get_sorted_quest_ids() const
-{
-    std::vector<QuestId> quest_ids;
-    std::transform(++this->quests.begin(), this->quests.end(), std::back_inserter(quest_ids), [](const auto &x) { return x.first; });
-    std::stable_sort(quest_ids.begin(), quest_ids.end(), [this](auto x, auto y) { return this->order_completed(x, y); });
-    return quest_ids;
-}
-
-bool QuestList::order_completed(QuestId id1, QuestId id2) const
-{
-    const auto &quest1 = this->get_quest(id1);
-    const auto &quest2 = this->get_quest(id2);
-    return (quest1.comptime != quest2.comptime) ? (quest1.comptime < quest2.comptime) : (quest1.level < quest2.level);
 }
 
 /*!
@@ -236,7 +80,7 @@ void determine_random_questor(CreatureEntity &creature, QuestType &quest)
  * @param q_ptr クエスト情報への参照ポインタ
  * @param stat ステータス(成功or失敗)
  */
-void record_quest_final_status(QuestType *q_ptr, PLAYER_LEVEL lev, QuestStatusType stat)
+void record_quest_final_status(QuestType *q_ptr, short lev, QuestStatusType stat)
 {
     q_ptr->status = stat;
     q_ptr->complev = lev;
