+++ src/effect/effect-monster-charm.cpp
@@ -1,6 +1,5 @@
 #include "effect/effect-monster-charm.h"
 #include "avatar/avatar.h"
-#include "dungeon/quest.h"
 #include "effect/effect-monster-util.h"
 #include "effect/spells-effect-util.h"
 #include "monster-floor/monster-remover.h"
@@ -19,6 +18,8 @@
 #include "spell/spells-diceroll.h"
 #include "status/bad-status-setter.h"
 #include "system/creature-entity.h"
+#include "system/dungeon/dungeon-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
