+++ src/load/dungeon-loader.cpp
@@ -1,5 +1,4 @@
 #include "load/dungeon-loader.h"
-#include "dungeon/quest.h"
 #include "floor/floor-save-util.h"
 #include "floor/floor-save.h"
 #include "load/dummy-loader.h"
@@ -7,6 +6,7 @@
 #include "load/load-util.h"
 #include "load/old/load-v1-5-0.h"
 #include "save/floor-writer.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/floor/floor-info.h"
 #include "system/monrace/monrace-definition.h"
 #include "util/bit-flags-calculator.h"
