+++ src/load/quest-loader.cpp
@@ -1,12 +1,13 @@
 #include "load/quest-loader.h"
 #include "artifact/fixed-art-types.h"
-#include "dungeon/quest.h"
 #include "load/angband-version-comparer.h"
 #include "load/load-util.h"
 #include "load/load-zangband.h"
 #include "load/savedata-old-flag-types.h"
 #include "object-enchant/trg-types.h"
+#include "system/angband-exceptions.h"
 #include "system/artifact-type-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/floor/floor-info.h"
 #include "system/floor/town-list.h"
 #include "system/monrace/monrace-definition.h"
