+++ src/monster-floor/one-monster-placer.cpp
@@ -38,6 +38,7 @@
 #include "player/player-sex.h"
 #include "player/player-status.h"
 #include "system/creature-entity.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
