+++ src/save/save.cpp
@@ -15,7 +15,6 @@
 #include "alliance/alliance.h"
 #include "artifact/fixed-art-types.h"
 #include "core/object-compressor.h"
-#include "dungeon/quest.h"
 #include "inventory/inventory-slot-types.h"
 #include "io/files-util.h"
 #include "io/report.h"
@@ -32,6 +31,7 @@
 #include "save/save-util.h"
 #include "system/artifact-type-definition.h"
 #include "system/creature-entity.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/floor/floor-info.h"
 #include "system/floor/town-list.h"
 #include "system/floor/wilderness-grid.h"
