+++ src/spell-kind/spells-floor.cpp
@@ -6,7 +6,6 @@
 
 #include "spell-kind/spells-floor.h"
 #include "action/travel-execution.h"
-#include "dungeon/quest.h"
 #include "flavor/flavor-describer.h"
 #include "flavor/object-flavor-types.h"
 #include "floor/floor-object.h"
@@ -27,6 +26,7 @@
 #include "system/artifact-type-definition.h"
 #include "system/creature-entity.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/terrain/terrain-tag.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
