+++ src/wizard/wizard-game-modifier.cpp
@@ -11,6 +11,7 @@
 #include "player-info/self-info.h"
 #include "system/building-type-definition.h"
 #include "system/creature-entity.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/dungeon/dungeon-id.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/floor/floor-info.h"
